jetpack tweaking
